ext4 crypto: don't try to skip ext4_init_crypto()
Currently we skip calling ext4_init_crypto() and taking the
'crypto_init' mutex if ext4_read_workqueue is non-NULL.  This is
incorrect because ext4_init_crypto() may not have completed by the time
ext4_read_workqueue is observed to be set, resulting in dereference of
NULL ext4_crypt_info_cachep.  A correct implementation of
"double-checked locking" would actually require that check be done of
the last pointer initialized, with memory barriers added.

For now just fix the bug in the trivial way by always taking the mutex.
There shouldn't be any noticeable overhead.

(This fix is also planned for upstream; this commit is just a quick fix
for this particular kernel.)

Bug: 65545897
Change-Id: I84ac62199b18f267c3d7e3be33df312e48e65463
Signed-off-by: Eric Biggers <ebiggers@google.com>