ARM64: configs: Don't build pronto WLAN for stock
I'm going to try to use the stock wlan module.
There are few reasons why I want to do so:
1. Xiaomi's wlan driver isn't 100% inline with CAF and
we don't have their sources for it. This introduces some issues
like wrong MAC address on stock.
2. Using stock kernel module implies that we can leave our /system
unmodified and make further development for stock much easier.

Biggest downside of this is that we can't be 100% sure that
Xiaomi will always pick all the recent security fixes from CAF.